requirements.txt
asgiref==3.5.2
astroid==2.12.13
dill==0.3.6
distlib==0.3.6
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.8.2
isort==5.11.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.6.0
pylint-plugin-utils==0.7
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
wrapt==1.14.1
